simplify API connection
