+++ services/types.ts
@@ -1,0 +1,1 @@
+export type State = "idle" | "loading" | "success" | "error";
